requirements.txt
markupsafe==2.0.1
Flask==1.1.4
Werkzeug==1.0.1
flask-sockets==0.2.1
gunicorn==20.1.0
gevent~=24.2.1
mysql-connector-python==8.0.26